- updated jarviscore to v0.7.79
